auth/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
loguru==0.7.3
pydantic==2.11.7
python-dotenv==1.1.0
python_bcrypt==0.3.2
python_jose==3.4.0
PyYAML==6.0.2
SQLAlchemy==2.0.41
pytest==8.4.0
pydantic[email]
python-multipart